[PATCH] blktrace: gcc-4 signed vs unsigned warning

+++ blktrace.c
@@ -755,7 +755,7 @@ int main(int argc, char *argv[])
 		fprintf(stderr,"%s does not appear to be a valid path\n",
 			relay_path);
 		return 1;
-	} else if (st.f_type != RELAYFS_TYPE) {
+	} else if (st.f_type != (long) RELAYFS_TYPE) {
 		fprintf(stderr,"%s does not appear to be a relay filesystem\n",
 			relay_path);
 		return 1;
